refactor: forward-compatibility for upcoming type improvements in Angular Pipes
There is an upcoming improvement in common pipes in Angular framework that would be released
in the next major version (v11). The changes can be found in
https://github.com/angular/angular/pull/37447.

This commit refactors the `CNCurrencyPipe` pipe to provide forward-compatibility for the next
major release of Angular framework. The improved typings make `transform` function signature
of the mentioned pipe incompatible with the `CurrencyPipe`. This commit updates the
`CNCurrencyPipe` pipe to initialize the `CurrencyPipe` in constructor instead of extending it
(to keep existing `transform` function signature) and use created instance inside the `transform`
function as needed.